requirements.txt
streamlit
pandas
matplotlib
PyYAML
bcrypt
streamlit-authenticator>=0.4.1
